backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
asyncpg==0.29.0
openai==1.3.5
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
sqlalchemy==2.0.23
psycopg2-binary
python-dotenv==1.0.0
model-context-protocol==0.2.0
google-generativeai==0.8.6